requirements.txt
google-generativeai
python-dotenv
schedule
google-auth
requests